Renaming filters package to chatfilters Renaming help package to cmdhelp
